Update
This program is a Implementation of the board game Chess, with a player opponent approach. All move types are supported, excluding en passant, castling and promotion(in development).
The application shows chess board in a browser with JS and also contains a C++ version which runs in the console. Players may interact and play a 2 player chess game. Technology stack: C++, JavaScript, Html, Css.